AP_Motors: convert to 32 bit motor mask

+++ libraries/AP_Motors/AP_MotorsHeli_Single.cpp
@@ -355,7 +355,7 @@ void AP_MotorsHeli_Single::calculate_scalars()
 
 // get_motor_mask - returns a bitmask of which outputs are being used for motors or servos (1 means being used)
 //  this can be used to ensure other pwm outputs (i.e. for servos) do not conflict
-uint16_t AP_MotorsHeli_Single::get_motor_mask()
+uint32_t AP_MotorsHeli_Single::get_motor_mask()
 {
     // heli uses channels 1,2,3,4 and 8
     // setup fast channels
